// src/app/models/repair-slip.ts
export interface RepairSlipApiResponse {
  message: string;
  totalItems: number;
  data: RepairSlip[];
}

/**
 * 修理モデル
 */
export interface RepairSlip {
  id: number;
  customer_type_division_id: number;
  store_id: number;
  client_id?: number;
  member_id?: number;
  last_name?: string;
  first_name?: string;
  last_name_kana?: string;
  first_name_kana?: string;
  tel?: string;
  mobile_number?: string;
  reception_date: string;
  reception_employee_id: number;
  status_division_id?: number;
  settle_status_division_id: number;
  incident_division_id?: number;
  remarks_1?: string;
  remarks_2?: string;
  created_at: string;
  created_id: number;
  updated_at?: string;
  updated_id?: number;
  deleted_at?: string;
  deleted_id?: number;
  store_name?: string;
  client_client_cd?: number;
  client_name?: string;
  client_name_kana?: string;
  client_pic_name?: string;
  client_pic_name_kana?: string;
  client_billing_cd?: number;
  client_postal_code?: string;
  client_province: string;
  client_locality: string;
  client_street_address: string;
  client_other_address: string;
  client_tel?: string;
  client_fax?: string;
  client_mail?: string;
  client_remarks_1?: string;
  client_remarks_2?: string;
  client_cutoff_date_billing?: number;
  client_scheduled_payment_date?: string;
  client_payment_division_id?: number;
  client_payment_term?: number;
  client_sales_tax_division_id?: number;
  client_division_sales_tax_code?: number;
  client_division_sales_tax_name?: string;
  client_division_sales_tax_value?: string;
  client_sales_tax_calc_division_id?: number;
  client_division_sales_tax_calc_code?: number;
  client_division_sales_tax_calc_name?: string;
  client_division_sales_tax_calc_value?: string;
  client_sales_fraction_division_id?: number;
  client_division_sales_fraction_code?: number;
  client_division_sales_fraction_name?: string;
  client_division_sales_fraction_value?: string;
  client_sales_tax_fraction_division_id?: number;
  client_division_sales_tax_fraction_code?: number;
  client_division_sales_tax_fraction_name?: string;
  client_division_sales_tax_fraction_value?: string;
  client_print_division_id?: number;
  client_division_print_code?: number;
  client_division_print_name?: string;
  client_division_print_value?: string;
  client_price_rank_id?: number;
  client_division_price_rank_code?: number;
  client_division_price_rank_name?: string;
  client_division_price_rank_value?: string;
  client_credit?: number;
  client_line_num?: number;
  client_title_division_id?: number;
  client_division_title_code?: number;
  client_division_title_name?: string;
  client_division_title_value?: string;
  client_custom_title?: string;
  member_name?: string;
  member_last_name?: string;
  member_last_name_kana?: string;
  member_first_name?: string;
  member_first_name_kana?: string;
  member_cd?: number;
  member_postal_code?: string;
  member_province: string;
  member_locality: string;
  member_street_address: string;
  member_other_address: string;
  member_tel?: string;
  member_mail?: string;
  member_point?: string;
  member_status_division_id?: number;
  member_division_status_code?: number;
  member_division_status_name?: string;
  member_division_status_value?: string;
  member_remarks_1?: string;
  member_remarks_2?: string;
  division_customer_type_name?: string;
  division_customer_type_code?: number;
  division_customer_type_value?: string;
  employee_reception_last_name?: string;
  employee_reception_first_name?: string;
  division_status_name: string;
  division_status_code: number;
  division_status_value: string;
  division_settle_status_name?: string;
  division_settle_status_code?: number;
  division_settle_status_value?: string;
  division_incident_name?: string;
  division_incident_code?: number;
  division_incident_value?: string;
  employee_created_last_name?: string;
  employee_created_first_name?: string;
  employee_updated_last_name?: string;
  employee_updated_first_name?: string;
  employee_deleted_last_name?: string;
  employee_deleted_first_name?: string;
}